Feature: Password Manager Entries
This branch introduces the Password Manager component in a form of Entries, the second main class of items alongside Vaults. The empty Secrets page in the application was replaced with the new view with the list of entries, where users may create, edit, view, and manage credentials containing usernames, emails, and passwords. Each entry icon is a 2x2 grid which in turn contains icons representing the types of information an entry stores, with separate icons for: email, username, password. The structure of the database was modified to store Vaults and Entries in separate directories ('secrets/vaults' and 'secrets/entries') in order to ensure groups (folders) of vaults are not displayed in the EntryListPage and vice versa.

List of changes:
- created details_hide.svg and details_show.svg icons and added them to app_icons.dart, they are used for buttons which toggle password visibility in entries
- created icon_entry_email.svg, icon_entry_password.svg and icon_entry_username.svg icons and added them to app_icons.dart, they are used to outline what sort of data each entry contains
- created entry_details_page.dart, containing read-only entry contents, with a support for copy-on-click for individual entry details
- created entry_details_editable_page.dart, containing editable entry contents, where a user may modify entry details
- created entry_page_type.dart enum, used to differentiate between creating a new entry or modifying an existing one
- created entry_entity.dart, entries_repository.dart, entries_service.dart and entry_model_factory.dart, adding infrastructure for Entries
- created entry_create_edit_status.dart enum, used to display the correct message in a dialog after successfully creating or modifying an entry
- created entry_list_page.dart and entry_list_page_cubit.dart, adding the list of entries view and defining their location as the 'secrets/entries' directory - separate directories for vaults and entries are created to ensure groups (folders) of vaults are not displayed in the EntryListPage (and vice versa)
- created entry_list_item.dart, defining the appearance of the entries on the list of entries view
- created entry_icon.dart, implementing Entry icons with their content preview
- updated vault_list_page.dart, redefining vaults location as the 'secrets/vaults' directory - separate directories for vaults and entries are created to ensure groups (folders) of entries are not displayed in the VaultListPage (and vice versa)
- updated group_model_factory.dart and a_secrets_model.dart, adding support for entries
- updated router.dart, adding new routes for Entries
- updated locator.dart, adding EntriesRepository, EntriesService and EntryModelFactory
- updated isar_database_manager.dart, adding EntryEntitySchema
- updated container_icon_grid.dart, container_icon_grid_item.dart and list_item_icon.dart, implementing support for Entry icons and their content preview
- updated group_icon.dart, network_group_icon.dart and vault_icon.dart with changes in container_icon_grid.dart
- updated app_service.dart, changing isDataBaseExist() method to search the 'secrets/vaults' and 'secrets/entries' directories for snggle files, instead of the root '/secrets' directory, as now it does not contain any data except for the automatically created 'vaults' and 'entries' directories and their secrets
- updated app_set_up_pin_page_cubit.dart to generate new 'secrets/vaults' and 'secrets/entries' directories and their secrets after first app PIN is created
- updated filesystem_storage_manager.dart, excluding directories 'secrets/vaults' and 'secrets/entries' from being deleted even if they are empty
- updated custom_text_field.dart, adding a bool to support both static (used in entry_details_page.dart and entry_details_editable_page.dart) and dynamically moving suffixes (used in wallet_create_page.dart), as well as adding onTap callback required for copying to work correctly when CustomTextField is wrapped in CopyWrapper
- updated wallet_create_page.dart, ensuring Solana derivation path keeps using dynamic suffix after changes in custom_text_field.dart
- updated copy_wrapper.dart, adding a callback required for copying to work correctly when CustomTextField is wrapped in CopyWrapper
- updated ethereum_transaction_details_page.dart, solana_transaction_details_page.dart, transaction_list_item_expansion.dart, wallet_details_page.dart, abi_chunks_list_item.dart and public_address_preview.dart with changes to copy_wrapper.dart

+++ lib/infra/managers/filesystem_storage/filesystem_storage_manager.dart
@@ -61,7 +61,8 @@ class FilesystemStorageManager {
 
     Directory parentDirectory = await _getParentDirectory(filesystemPath);
     bool parentDirectoryEmptyBool = parentDirectory.listSync().isEmpty;
-    if (parentDirectoryEmptyBool) {
+    bool parentDirectoryRootBool = parentDirectory.path.endsWith('/vaults') || parentDirectory.path.endsWith('/entries');
+    if (parentDirectoryEmptyBool && parentDirectoryRootBool == false) {
       await parentDirectory.delete();
     }
   }
